Created code_of_conduct.md
Automatically accepted by everyone committing.
Instances of abusive, harassing, or otherwise unacceptable behavior may be reported by contacting the project team at Riverssen@gmail.com. The project team will review and investigate all complaints, and will respond in a way that it deems appropriate to the circumstances. The project team is obligated to maintain confidentiality with regard to the reporter of an incident. Further details of specific enforcement policies may be posted separately.
Project maintainers who do not follow or enforce the Code of Conduct in good faith may face temporary or permanent repercussions as determined by other members of the project's leadership.